Added tests and new example
